# openapi.yaml
openapi: 3.1.0
info:
  title: PromptHub Enterprise
  description: Stand-Alone Enterprise AI Prompt Library, Engineering, Testing & Governance
    Platform
  version: 1.0.0
paths:
...
  /api/v1/admin/users:
    get:
      tags:
      - admin
      summary: List Users
      operationId: list_users_api_v1_admin_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserSummary'
                type: array
                title: Response List Users Api V1 Admin Users Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - admin
      summary: Create User
      operationId: create_user_api_v1_admin_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/users/{user_id}:
    put:
      tags:
      - admin
      summary: Update User
      operationId: update_user_api_v1_admin_users__user_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Delete User
      operationId: delete_user_api_v1_admin_users__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete User Api V1 Admin Users  User Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserCreate:
      properties:
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
        email:
          type: string
          title: Email
          default: ''
        department:
          type: string
          title: Department
          default: ''
        title:
          type: string
          title: Title
          default: ''
        role:
          type: string
          title: Role
          default: USER
        password:
          type: string
          title: Password
          default: ''
      type: object
      required:
      - username
      - display_name
      title: UserCreate
    UserSummary:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
        email:
          type: string
          title: Email
          default: ''
        role:
          type: string
          title: Role
        department:
          type: string
          title: Department
          default: ''
        title:
          type: string
          title: Title
          default: ''
        is_active:
          type: boolean
          title: Is Active
          default: true
      type: object
      required:
      - id
      - user_id
      - username
      - display_name
      - role
      title: UserSummary
    UserUpdate:
      properties:
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer